requirements.txt
bcrypt==4.1.3
certifi==2024.2.2
Flask==3.0.3
flask-cors==4.0.1
google-auth==2.29.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
pillow==10.3.0
